VariBlock: Add static cut policy
Add a StaticCutPolicy class and its indexer to the VariBlockPolicies.chpl
file. For each dimension, user of the policy can choose how many and how large
chunks are to be produced in that particular dimension. Word static in name
implies that cuts must be set before domain map is created. Setting cuts
after creation of domain map will not change how domain map distributes data,
and in certain situations causes the program to fail.